reproject_depthmap/requirements.txt
Pillow
matplotlib
numpy
open3d>=0.15.1
torch>=1.8.1
